pyproject.toml
[project]
name = "pdf-toolkit"
version = "0.1.0"
description = "A collection of pdf utility tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "pymupdf>=1.26.7",
]
